requirements.txt
Pillow==9.3.0
torch==2.1.0
torchvision==0.16.0
scipy==1.11.3